pyproject.toml
[project]
name = "yt-hist"
version = "0.1.0"
description = "Local YouTube history semantic search and RAG service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "fastembed>=0.3.0",
  "llama-index-core>=0.11.0",
  "llama-index-embeddings-fastembed>=0.2.0",
]

[project.optional-dependencies]
openai = [
  "llama-index-llms-openai>=0.2.0",
]

[project.scripts]
ythist = "ythist.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ythist"]
